[MIG] 10.0 port project_task_default_stage (#201)
