+++ project_task_default_stage/README.rst
@@ -7,7 +7,7 @@ Project Task Default Stage
 ==========================
 
 This module restores the ``case_default`` fields to Project Tasks and project
-stage tab, removed in Odoo 9.0.
+stage tab, removed in Odoo 9.0 and later.
 
 Configuration
 =============
@@ -19,15 +19,14 @@ Odoo 8.0.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/140/9.0
+   :target: https://runbot.odoo-community.org/runbot/140/10.0
 
 Bug Tracker
 ===========
 
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/project-service/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
-`here <https://github.com/OCA/project-service/issues/new?body=module:%20project_case_default%0Aversion:%209.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+If you spotted it first, help us smashing it by providing a detailed and welcomed feedback.
 
 Credits
 =======
