+++ project_task_default_stage/__manifest__.py
@@ -6,9 +6,9 @@
 {
     'name': 'Project Task Default Stage',
     'summary': 'Recovery default task stage projects from v8',
-    'version': '9.0.1.0.0',
+    'version': '10.0.1.0.0',
     'category': 'Project',
-    'author': 'Incaser Informatica S.L., '
+    'author': 'Tecnativa, '
               'Odoo Community Association (OCA)',
     'license': 'AGPL-3',
     'depends': [
@@ -18,5 +18,5 @@
         'views/project_view.xml',
         'data/project_data.xml',
     ],
-    'installable': False,
+    'installable': True,
 }
