+++ project_task_default_stage/data/project_data.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data noupdate="1">
+<odoo noupdate="1">
 
     <!-- Task Stages -->
 
@@ -48,5 +47,4 @@
         <field name="fold" eval="True"/>
     </record>
 
-</data>
-</openerp>
+</odoo>
