--- project_task_default_stage/i18n/project_case_default.pot
@@ -1,92 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* project_task_default_stage
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 9.0c\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-12-29 10:35+0000\n"
-"PO-Revision-Date: 2015-12-29 10:35+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_analysis
-msgid "Analysis"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_cancel
-msgid "Cancelled"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:ir.model.fields,help:project_task_default_stage.field_project_project_use_timesheets
-msgid "Check this field if this project manages timesheets"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:ir.model.fields,field_description:project_task_default_stage.field_project_task_type_case_default
-msgid "Default for New Projects"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_design
-msgid "Design"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_development
-msgid "Development"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_deployment
-msgid "Done"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:ir.model.fields,help:project_task_default_stage.field_project_task_type_case_default
-msgid "If you check this field, this stage will be proposed by default on each new project. It will not assign this stage to existing projects."
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_merge
-msgid "Merge"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:ir.model,name:project_task_default_stage.model_project_project
-msgid "Project"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:ir.ui.view,arch_db:project_task_default_stage.edit_project
-msgid "Project Stages"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_specification
-msgid "Specification"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:ir.model,name:project_task_default_stage.model_project_task_type
-msgid "Task Stage"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:project.task.type,name:project_task_default_stage.project_tt_testing
-msgid "Testing"
-msgstr ""
-
-#. module: project_task_default_stage
-#: model:ir.model.fields,field_description:project_task_default_stage.field_project_project_use_timesheets
-msgid "Timesheets"
-msgstr ""
-
