+++ project_task_default_stage/models/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
+from . import project_task_type
 from . import project
