+++ project_task_default_stage/models/project_task_type.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# (c) 2015 Incaser Informatica S.L. - Sergio Teruel
+# (c) 2015 Incaser Informatica S.L. - Carlos Dauden
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from odoo import models, fields
+
+
+class ProjectTaskType(models.Model):
+    _inherit = 'project.task.type'
+
+    case_default = fields.Boolean(
+        string='Default for New Projects',
+        help='If you check this field, this stage will be proposed by default '
+             'on each new project. It will not assign this stage to existing '
+             'projects.')
