+++ project_task_default_stage/tests/test_project_task_default_stage.py
@@ -3,7 +3,7 @@
 # (c) 2015 Incaser Informatica S.L. - Carlos Dauden
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from openerp.tests.common import TransactionCase
+from odoo.tests.common import TransactionCase
 
 
 class TestProjectCaseDefault(TransactionCase):
