+++ project_task_default_stage/views/project_view.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
+<odoo>
 
     <record id="task_type_edit" model="ir.ui.view">
         <field name="name">project.task.type.form.inherit</field>
@@ -28,5 +27,4 @@
         </field>
     </record>
 
-</data>
-</openerp>
+</odoo>
